requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
alembic==1.4.3
antlr4-python3-runtime==4.8
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
blis==0.7.4
cachetools==4.2.0
catalogue==1.0.0
category-encoders==2.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cliff==3.5.0
cmaes==0.7.0
cmd2==1.4.0
colorama==0.4.4
colorlog==4.6.2
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
defusedxml==0.6.0
en-core-web-sm==2.3.1
entrypoints==0.3
extractcontent3==0.0.2
filelock==3.0.12
fsspec==0.8.5
future==0.18.2
gensim==3.8.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
hydra-core==1.0.5
idna==2.10
importlib-resources==5.0.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lightgbm==3.1.1
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
ml-metrics==0.1.4
more-itertools==8.6.0
multidict==5.1.0
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.5
notebook==6.1.6
numpy==1.19.5
nyaggle==0.1.4
oauthlib==3.1.0
omegaconf==2.0.5
optuna==2.3.0
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
patsy==0.5.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plac==1.1.3
plotly==4.14.2
preshed==3.0.5
prettytable==0.7.2
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pyperclip==1.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-lightning==1.1.3
pytz==2020.5
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7
sacremoses==0.0.43
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
signate==0.9.7
six==1.15.0
sklearn==0.0
smart-open==4.1.0
spacy==2.3.5
SQLAlchemy==1.3.22
srsly==1.0.5
statsmodels==0.12.1
stevedore==3.3.0
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
terminado==0.9.1
testpath==0.4.4
texthero==1.0.9
thinc==7.4.5
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.7.1
torchtext==0.8.1
tornado==6.1
tqdm==4.55.2
traitlets==5.0.5
transformers==4.1.1
typing-extensions==3.7.4.3
Unidecode==1.1.2
urllib3==1.26.2
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
wordcloud==1.8.1
xfeat==0.1.1
yarl==1.6.3
